+++ packages/orq-rc/src/models/operations/listdatasets.ts
@@ -156,7 +156,7 @@ export const ListDatasetsData$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-02-23T11:02:22.537Z",
+    "2026-02-23T12:22:02.215Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {